requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.5
django-filter==24.3
django-pytest==0.2.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
